+++ spec/lib/rails_ai_bridge/version_spec.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require "spec_helper"
+
+RSpec.describe "RailsAiBridge::VERSION" do
+  specify do
+    expect(RailsAiBridge::VERSION).to be_a(String)
+    expect(RailsAiBridge::VERSION).to match(/\A\d+\.\d+\.\d+/)
+  end
+end